Separando aplicação Laravel na pasta backend

+++ .docker/entrypoint.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
 
 #On error no such file entrypoint.sh, execute in terminal - dos2unix .docker\entrypoint.sh
-cp .env.example .env
-cp .env.testing.example .env.testing
+cd backend
+if [ ! -f ".env" ]; then
+  cp .env.example .env
+fi
+if [ ! -f ".env.testing" ]; then
+  cp .env.testing.example .env.testing
+fi
 chown -R www-data:www-data .
 composer install
 php artisan key:generate
